requirements_pip.txt
Package                       Version
----------------------------- -------
backports.functools-lru-cache 1.6.4
black                         22.3.0
boto3                         1.21.37
Bottleneck                    1.3.4
brotlipy                      0.7.0
dataclasses                   0.8
findspark                     2.0.1
flake8                        4.0.1
jupyter-book                  0.12.3
numexpr                       2.8.0
pandas                        1.3.4
pip                           22.0.4
progressbar                   2.5
pyarrow                       0.11.1
pyOpenSSL                     22.0.0
PySocks                       1.7.1
pyspark                       3.2.1
pytest-cov                    3.0.0
python-dotenv                 0.20.0
toml                          0.10.2
tqdm                          4.64.0
yamlordereddictloader         0.4.0
Package                       Version
----------------------------- -------
backports.functools-lru-cache 1.6.4
black                         22.3.0
boto3                         1.21.37
Bottleneck                    1.3.4
brotlipy                      0.7.0
dataclasses                   0.8
findspark                     2.0.1
flake8                        4.0.1
jupyter-book                  0.12.3
numexpr                       2.8.0
pandas                        1.3.4
pip                           22.0.4
progressbar                   2.5
pyarrow                       0.11.1
pyOpenSSL                     22.0.0
PySocks                       1.7.1
pyspark                       3.2.1
pytest-cov                    3.0.0
python-dotenv                 0.20.0
toml                          0.10.2
tqdm                          4.64.0
yamlordereddictloader         0.4.0
